chore: utils 폴더 정리
home(+exam-info), common 폴더 추가
